backend/pyproject.toml
[project]
name = "practice-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "argon2-cffi>=23.1.0",
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.1",
]
